Classify RTags as inactive if it can't provide compilation flags
When working on a project, it's quite common to also open source files
outside of the current project. Trying to analyze those files will
inevitably fail because RTags can't provide compilation flags for them.

This modifies `flycheck-clang-analyzer--rtags-active' to consider RTags to
be inactive if it can't provide compilation flags for the current buffer,
even if it's enabled and running.